refactor(memory): share atomicWrite; derive KINDS; reuse storeHint; dedup lock backoff
Behavior-preserving structural cleanups for mauto memory Slice 1: extract
the duplicated atomic-write algorithm from MemoryStore/ResultStore into
src/util/atomic.js, derive the memory KINDS list from FILES and colocate
HEADERS in src/memory/paths.js, reuse storeHint() for the memory-warnings
hint in handleResultFinalize, and hoist lock.js's duplicated
backoff-or-throw block into a single closure. Full suite (551 tests)
stays green with no test-expectation changes.
