+++ src/memory/lock.js
@@ -22,6 +22,12 @@ function sleepSync(ms) {
 function acquire(lockPath, { now = Date.now, sleep = sleepSync } = {}) {
   fs.mkdirSync(path.dirname(lockPath), { recursive: true });
   const deadline = now() + MAX_WAIT_MS;
+  const backoffOrThrow = () => {
+    if (now() >= deadline) {
+      throw new Error(`could not acquire memory lock at ${lockPath} within ${MAX_WAIT_MS}ms`);
+    }
+    sleep(RETRY_MS);
+  };
   for (;;) {
     try {
       const fd = fs.openSync(lockPath, 'wx');
@@ -42,22 +48,12 @@ function acquire(lockPath, { now = Date.now, sleep = sleepSync } = {}) {
           // Not the benign vanished-lock case — apply the bounded backoff so
           // a persistent stat/unlink failure (EACCES/EPERM/EBUSY, ...) times
           // out instead of busy-looping the CPU forever.
-          if (now() >= deadline) {
-            throw new Error(
-              `could not acquire memory lock at ${lockPath} within ${MAX_WAIT_MS}ms`
-            );
-          }
-          sleep(RETRY_MS);
+          backoffOrThrow();
         }
         continue; // ENOENT: lock vanished between open and stat — retry now;
         // any other error already backed off above before retrying.
       }
-      if (now() >= deadline) {
-        throw new Error(
-          `could not acquire memory lock at ${lockPath} within ${MAX_WAIT_MS}ms`
-        );
-      }
-      sleep(RETRY_MS);
+      backoffOrThrow();
     }
   }
 }
